backend/pyproject.toml
[project]
name = "be-coding-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.31",
]

[dependency-groups]
dev = [
    "faker>=40.1.2",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=html --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
exclude = ["alembic/versions/*.py"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
